% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecrt_sim.R
\name{fecrt_sim}
\alias{fecrt_sim}
\title{Title}
\usage{
fecrt_sim(n_animals)
}
\arguments{
\item{n_animals}{the number of animals to simulate}
}
\description{
Title
}
\details{
Allows for:
\itemize{
\item treatment/control or pre/post treatment
\item sample size of animals
\item pools of P animals
\item pasture collections
}
}
