% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecrtsim-package.R
\docType{package}
\name{fecrtsim-package}
\alias{fecrtsim-package}
\alias{fecrtsim}
\title{What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
Some description of the package
}
\examples{
## An example function:
df <- example_function()
df

## To cite this package in publications use:
citation("fecrtsim")

}
\references{
Some article that you might want users to look at
}
\seealso{
\code{\link[neatpkg]{pkg_new}} for the function that created this package
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Other contributors:
\itemize{
  \item Mackenzie Smith \email{Mackenzie.Smith@uky.edu} [contributor]
}

}
\keyword{internal}
