% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_defecation.R
\name{simulate_defecation}
\alias{simulate_defecation}
\alias{group_defecation}
\title{Simulate defecation events}
\usage{
simulate_defecation(
  rho_0 = 1.2,
  loghaz_0 = -5.6,
  rho_1 = 2.6,
  loghaz_1 = -13,
  N = 1L,
  end_time = 60 * 3
)

group_defecation(defecations, breaks = seq(0, 60 * 3, by = 60))
}
\arguments{
\item{rho_0}{the rho (shape) parameter for the time to first defecation event}

\item{loghaz_0}{the log hazard of first defecation event (see details for relationship to the mean and scale parameter)}

\item{rho_1}{the rho (shape) parameter for the time between subsequent defecation events}

\item{loghaz_1}{the log hazard of subsequent defecation events (see details for relationship to the mean and scale parameter)}

\item{N}{the number of animals to simulate}

\item{end_time}{the time period for the simulation (presumed to be in minutes, although this is implicitly on the same scale as the rho/lambda parameters above)}
}
\value{
a tibble with Animal (factor, with levels giving all animals including those with implicitly zero defecation events), Time (time of defecation event), and Number (order of defecation events for that animal)
}
\description{
Simulate defecation events
}
\details{
Different rho/loghaz parameters for each animal can be supplied with a vector of length equal to N (length-1 inputs will be recycled, but no other recycling is performed).

Note that the inputs for the time-to-event distributions are provided as rho and lambda:  this matches the parameterisation used by JAGS, but not that used by R.  The relationship between parameters is:
shape = rho
lambda = exp(log_hazard)
scale = lambda^(-1/rho)
mean = scale * gamma(1 + 1/shape)
lambda = (log(mean) – loggam(1 + 1/rho)) * -rho
(Where gamma() denotes the Gamma function, and loggam() is the Log Gamma function)

The reason for using this parameterisation is that animal-level effects on log-hazard can be more easily specified by adding e.g. normally distributed random effects to loghaz_0 and loghaz_1
}
\examples{
defecations <- simulate_defecation(N=1000)

}
