% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecrt_sim.R
\name{fecrt_sim}
\alias{fecrt_sim}
\title{Title}
\usage{
fecrt_sim(n_animals)
}
\arguments{
\item{n_sim}{the number of FECRT to simulate}

\item{design}{either paired or unpaired}

\item{n_treatment}{the number of animals in the treatment group}

\item{n_control}{the number of animals in the control group (ignored for paired design)}

\item{n_pools}{the number of pooled samples}
}
\description{
Title
}
\details{
Allows for:
\itemize{
\item treatment/control or pre/post treatment
\item sample size of animals
\item pools of P animals
\item pasture collections
}
}
